chore: refactor integration test helper classes
