+++ src/itest/kotlin/nl/lifely/zac/itest/AppContainerTest.kt
@@ -8,16 +8,16 @@ package nl.lifely.zac.itest
 import io.kotest.assertions.json.shouldContainJsonKeyValue
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
-import nl.lifely.zac.itest.client.ZacClient
+import nl.lifely.zac.itest.client.ItestHttpClient
 import nl.lifely.zac.itest.config.ItestConfiguration.ZAC_MANAGEMENT_URI
 
-private val zacClient: ZacClient = ZacClient()
+private val itestHttpClient = ItestHttpClient()
 
 class AppContainerTest : BehaviorSpec({
     given("ZAC Docker container and all related Docker containers are running") {
         When("the health endpoint is called") {
             then("the response should be ok and the status should be UP") {
-                zacClient.performGetRequest(
+                itestHttpClient.performGetRequest(
                     url = "${ZAC_MANAGEMENT_URI}/health"
                 ).use { response ->
                     response.isSuccessful shouldBe true
@@ -31,7 +31,7 @@ class AppContainerTest : BehaviorSpec({
     given("ZAC Docker container and all related Docker containers are running") {
         When("the metrics endpoint is called") {
             then("the response should be ok and the the uptime var should be present") {
-                zacClient.performGetRequest(
+                itestHttpClient.performGetRequest(
                     url = "${ZAC_MANAGEMENT_URI}/metrics"
                 ).use { response ->
                     response.isSuccessful shouldBe true
